broke script.js into these components
