+++ src/nodeobject/Goal.js
@@ -1,0 +1,6 @@
+export class Goal {
+    constructor(name) {
+        this.name = name;
+    }
+}
+
