+++ src/nodeobject/script.js
@@ -1,81 +1,65 @@
-function createNote(key, content, date) {
-    //Used later for clarity
-    let newIndex;
-    //Creates our empty object
-    let noteJSON = {};
-    //Done to save a little space, and it means you can still create one note/second
-    let key = Math.floor(Date.now()/1000); 
-    noteJSON.key = key;
-    //Set the content using the parameters
-    noteJSON.cnt = content;
-    noteJSON.dte = date;
-    //Creates an empty array we can add subcontent to later
-    noteJSON.sub = [];
-    //Stores the JSON in string form in localStorage, using the creation time as a key
-    localStorage.setItem(`${key}`, JSON.stringify(noteJSON));
-    //Checks if the date index exists, and inserts the note into the date index in localStorage
-    let dateIndex = localStorage.getItem(`${date}`);
-    if(dateIndex == null){
-        localStorage.setItem(`${date}`, JSON.stringify([key]))
-    } else {
-        newIndex = JSON.parse(dateIndex).push(key);
-        localStorage.setItem(`${date}`, JSON.stringify(newIndex));
-    }
-    //Returns the noteJSON to easily pass into the HTML element rendering function
-    return noteJSON;
+import { Note } from './Note.js';
+import { Task } from './Task.js';
+import { Goal } from './Goal.js';
+
+function createNote(content, date) {
+    return new Note(date, content);
+}
+// function store(thingToStore){
+//     if (thingstore instanceof note){
+//         storeNote(thing)
+//     }
+// }
+function storeNote(note){
+    // localStorage.setItem(`${key}`, JSON.stringify(noteJSON));
+    // //Checks if the date index exists, and inserts the note into the date index in localStorage
+    // let dateIndex = localStorage.getItem(`${date}`);
+    // if(dateIndex == null){
+    //     localStorage.setItem(`${date}`, JSON.stringify([key]))
+    // } else {
+    //     newIndex = JSON.parse(dateIndex).push(key);
+    //     localStorage.setItem(`${date}`, JSON.stringify(newIndex));
+    // }
+    // //Returns the noteJSON to easily pass into the HTML element rendering function
+    // return noteJSON;
 }
 
-function createTask(key, content, date, time, goal, priority) {
-    //Used later for clarity
-    let newIndex;
-    //Creates our empty object
-    let taskJSON = {};
-    //Done to save a little space, and it means you can still create one note/second
-    let key = Math.floor(Date.now()/1000); 
-    taskJSON.key = key;
-    //Set the content using the parameters
-    taskJSON.cnt = content;
-    taskJSON.dte = date;
-    taskJSON.tim = time;
-    taskJSON.gol = goal;
-    taskJSON.prt = priority;
-    //Tracks if the goal is complete
-    taskJSON.cmp = false; 
-    //Creates an empty array we can add subcontent to later
-    taskJSON.sub = [];
-    //Stores the JSON in string form in localStorage, using the creation time as a key
-    localStorage.setItem(`${key}`, JSON.stringify(taskJSON));
-    //Checks if the date index exists, and inserts the task into the date index in localStorage
-    let dateIndex = localStorage.getItem(`${date}`);
-    if(dateIndex == null){
-        localStorage.setItem(`${date}`, JSON.stringify([key]))
-    } else {
-        newIndex = JSON.parse(dateIndex).push(key);
-        localStorage.setItem(`${date}`, JSON.stringify(newIndex));
-    }
-    //Inserts the task into the goal index in localStorage
-    newIndex = JSON.parse(localStorage.getItem(`${goal}`)).push(key);
-    localStorage.setItem(`${goal}`, newIndex);
-    //Checks if the priority index exists, and inserts the task into the priority index in localStorage
-    priorityIndex = localStorage.getItem(`${priority}`);
-    if(priorityIndex == null){
-        localStorage.setItem(`${priority}`, JSON.stringify([key]))
-    } else {
-        newIndex = JSON.parse(priorityIndex).push(key);
-        localStorage.setItem(`${priority}`, JSON.stringify(newIndex));
-    }
-    //Returns the noteJSON to easily pass into the HTML element rendering function
-    return taskJSON;
+function createTask(date, content, subTasks, time, goal, priority, isComplete) {
+    return new Task(date, content, subTasks, time, goal, priority, isComplete);
+}
+
+function storeTask(task){
+    // localStorage.setItem(`${key}`, JSON.stringify(taskJSON));
+    // //Checks if the date index exists, and inserts the task into the date index in localStorage
+    // let dateIndex = localStorage.getItem(`${date}`);
+    // if(dateIndex == null){
+    //     localStorage.setItem(`${date}`, JSON.stringify([key]))
+    // } else {
+    //     newIndex = JSON.parse(dateIndex).push(key);
+    //     localStorage.setItem(`${date}`, JSON.stringify(newIndex));
+    // }
+    // //Inserts the task into the goal index in localStorage
+    // newIndex = JSON.parse(localStorage.getItem(`${goal}`)).push(key);
+    // localStorage.setItem(`${goal}`, newIndex);
+    // //Checks if the priority index exists, and inserts the task into the priority index in localStorage
+    // priorityIndex = localStorage.getItem(`${priority}`);
+    // if(priorityIndex == null){
+    //     localStorage.setItem(`${priority}`, JSON.stringify([key]))
+    // } else {
+    //     newIndex = JSON.parse(priorityIndex).push(key);
+    //     localStorage.setItem(`${priority}`, JSON.stringify(newIndex));
+    // }
+    // //Returns the noteJSON to easily pass into the HTML element rendering function
+    // return taskJSON;
 }
 
 function createGoal(name) {
-    //Checks if a goal by that name already exists, if so, return 1
-    if(localStorage.getItem(name) != null){
-        return 1;
-    }
-    localStorage.setItem(name, JSON.stringify([]));
+    return new Goal(name);
 }
 
+function storeGoal(goal){
+    //localStorage.setItem(name, JSON.stringify([]));
+}
 function editNote(key, content, date) {
     let noteJSON = JSON.parse(localStorage.getItem(key));
     //Edit content if necessary
